# apps/api/openapi/v1.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Prescriptive Maintenance API",
    "summary": "Contrato v1 de análise e ciclo documental.",
    "description": "Contrato público v1 com runtime de análise selecionado explicitamente por configuração. O cabeçalho X-Analysis-Mode informa somente o modo sanitizado em uso.",
    "version": "1.0.0"
  },
  "paths": {
...
    "/documents": {
      "get": {
        "tags": [
          "documents"
        ],
        "summary": "Lista documentos e estados do ciclo",
        "operationId": "listDocuments",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DocumentListResponse"
                }
              }
            },
            "headers": {
              "X-Analysis-Mode": {
                "$ref": "#/components/headers/AnalysisMode"
              }
            }
          },
          "503": {
            "description": "O ciclo documental está temporariamente indisponível.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "headers": {
              "X-Analysis-Mode": {
                "$ref": "#/components/headers/AnalysisMode"
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "documents"
        ],
        "summary": "Registra metadados de um documento",
        "operationId": "registerDocument",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RegisterDocumentRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Registro recebido; nunca implica aprovação.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ReceivedDocument"
                },
                "example": {
                  "document_id": "doc_bbbbbbbbbbbb",
                  "filename": "inspection.synthetic.pdf",
                  "media_type": "application/pdf",
                  "size_bytes": 2048,
                  "sha256": "bbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbb",
                  "created_at": "2030-01-02T03:04:05Z",
                  "updated_at": "2030-01-02T03:04:05Z",
                  "status": "received"
                }
              }
            },
            "headers": {
              "X-Analysis-Mode": {
                "$ref": "#/components/headers/AnalysisMode"
              }
            }
          },
          "409": {
            "description": "O comando documental conflita com o estado armazenado.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "headers": {
              "X-Analysis-Mode": {
                "$ref": "#/components/headers/AnalysisMode"
              }
            }
          },
          "422": {
            "description": "Requisição inválida; nenhuma porta interna é chamada.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "headers": {
              "X-Analysis-Mode": {
                "$ref": "#/components/headers/AnalysisMode"
              }
            }
          },
          "503": {
            "description": "O ciclo documental está temporariamente indisponível.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "headers": {
              "X-Analysis-Mode": {
                "$ref": "#/components/headers/AnalysisMode"
              }
            }
          }
        }
      }
    },
    "/documents/{document_id}": {
      "get": {
        "tags": [
          "documents"
        ],
        "summary": "Consulta um documento",
        "operationId": "getDocument",
        "parameters": [
          {
            "name": "document_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "pattern": "^doc_[a-z0-9_]{3,64}$",
              "title": "Document Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Documento sintético em um estado válido.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DocumentResponse"
                },
                "examples": {
                  "failed": {
                    "summary": "Documento sintético failed",
                    "value": {
                      "document_id": "doc_synthetic_failed",
                      "filename": "failed.synthetic.pdf",
                      "media_type": "application/pdf",
                      "size_bytes": 1024,
                      "sha256": "aaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa",
                      "created_at": "2030-01-02T03:04:05Z",
                      "updated_at": "2030-01-02T03:04:05Z",
                      "status": "failed",
                      "decision_note": null,
                      "failure": {
                        "code": "synthetic_processing_failure",
                        "message": "Falha controlada do fake sintético."
                      },
                      "superseded_by_document_id": null
                    }
                  },
                  "approved": {
                    "summary": "Documento sintético approved",
                    "value": {
                      "document_id": "doc_synthetic_manual",
                      "filename": "manual.synthetic.pdf",
                      "media_type": "application/pdf",
                      "size_bytes": 1024,
                      "sha256": "aaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa",
                      "created_at": "2030-01-02T03:04:05Z",
                      "updated_at": "2030-01-02T03:04:05Z",
                      "status": "approved",
                      "decision_note": "Conteúdo inteiramente sintético aprovado para o fake.",
                      "failure": null,
                      "superseded_by_document_id": null
                    }
                  },
                  "pending_approval": {
                    "summary": "Documento sintético pending_approval",
                    "value": {
                      "document_id": "doc_synthetic_pending",
                      "filename": "pending.synthetic.pdf",
                      "media_type": "application/pdf",
                      "size_bytes": 1024,
                      "sha256": "aaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa",
                      "created_at": "2030-01-02T03:04:05Z",
                      "updated_at": "2030-01-02T03:04:05Z",
                      "status": "pending_approval",
                      "decision_note": null,
                      "failure": null,
                      "superseded_by_document_id": null
                    }
                  },
                  "processing": {
                    "summary": "Documento sintético processing",
                    "value": {
                      "document_id": "doc_synthetic_processing",
                      "filename": "processing.synthetic.pdf",
                      "media_type": "application/pdf",
                      "size_bytes": 1024,
                      "sha256": "aaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa",
                      "created_at": "2030-01-02T03:04:05Z",
                      "updated_at": "2030-01-02T03:04:05Z",
                      "status": "processing",
                      "decision_note": null,
                      "failure": null,
                      "superseded_by_document_id": null
                    }
                  },
                  "received": {
                    "summary": "Documento sintético received",
                    "value": {
                      "document_id": "doc_synthetic_received",
                      "filename": "received.synthetic.pdf",
                      "media_type": "application/pdf",
                      "size_bytes": 1024,
                      "sha256": "aaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa",
                      "created_at": "2030-01-02T03:04:05Z",
                      "updated_at": "2030-01-02T03:04:05Z",
                      "status": "received",
                      "decision_note": null,
                      "failure": null,
                      "superseded_by_document_id": null
                    }
                  },
                  "rejected": {
                    "summary": "Documento sintético rejected",
                    "value": {
                      "document_id": "doc_synthetic_rejected",
                      "filename": "rejected.synthetic.pdf",
                      "media_type": "application/pdf",
                      "size_bytes": 1024,
                      "sha256": "aaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa",
                      "created_at": "2030-01-02T03:04:05Z",
                      "updated_at": "2030-01-02T03:04:05Z",
                      "status": "rejected",
                      "decision_note": "Metadados sintéticos incompletos.",
                      "failure": null,
                      "superseded_by_document_id": null
                    }
                  },
                  "superseded": {
                    "summary": "Documento sintético superseded",
                    "value": {
                      "document_id": "doc_synthetic_superseded",
                      "filename": "old.synthetic.pdf",
                      "media_type": "application/pdf",
                      "size_bytes": 1024,
                      "sha256": "aaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa",
                      "created_at": "2030-01-02T03:04:05Z",
                      "updated_at": "2030-01-02T03:04:05Z",
                      "status": "superseded",
                      "decision_note": null,
                      "failure": null,
                      "superseded_by_document_id": "doc_synthetic_manual"
                    }
                  }
                }
              }
            },
            "headers": {
              "X-Analysis-Mode": {
                "$ref": "#/components/headers/AnalysisMode"
              }
            }
          },
          "404": {
            "description": "Recurso não encontrado.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "headers": {
              "X-Analysis-Mode": {
                "$ref": "#/components/headers/AnalysisMode"
              }
            }
          },
          "422": {
            "description": "Requisição inválida; nenhuma porta interna é chamada.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "headers": {
              "X-Analysis-Mode": {
                "$ref": "#/components/headers/AnalysisMode"
              }
            }
          },
          "503": {
            "description": "O ciclo documental está temporariamente indisponível.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "headers": {
              "X-Analysis-Mode": {
                "$ref": "#/components/headers/AnalysisMode"
              }
            }
          }
        }
      }
    },
    "/documents/{document_id}/approve": {
      "post": {
        "tags": [
          "documents"
        ],
        "summary": "Aprova um documento pendente",
        "operationId": "approveDocument",
        "parameters": [
          {
            "name": "document_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "pattern": "^doc_[a-z0-9_]{3,64}$",
              "title": "Document Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ApproveDocumentRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApprovedDocument"
                }
              }
            },
            "headers": {
              "X-Analysis-Mode": {
                "$ref": "#/components/headers/AnalysisMode"
              }
            }
          },
          "404": {
            "description": "Recurso não encontrado.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "headers": {
              "X-Analysis-Mode": {
                "$ref": "#/components/headers/AnalysisMode"
              }
            }
          },
          "409": {
            "description": "O comando documental conflita com o estado armazenado ou a transição solicitada não é válida para o estado atual.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "headers": {
              "X-Analysis-Mode": {
                "$ref": "#/components/headers/AnalysisMode"
              }
            }
          },
          "422": {
            "description": "Requisição inválida; nenhuma porta interna é chamada.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "headers": {
              "X-Analysis-Mode": {
                "$ref": "#/components/headers/AnalysisMode"
              }
            }
          },
          "503": {
            "description": "O ciclo documental está temporariamente indisponível.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "headers": {
              "X-Analysis-Mode": {
                "$ref": "#/components/headers/AnalysisMode"
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "ApproveDocumentRequest": {
        "properties": {
          "note": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 500,
                "minLength": 1
              },
              {
                "type": "null"
              }
            ],
            "title": "Note"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "title": "ApproveDocumentRequest"
      },
...
      "DocumentListResponse": {
        "properties": {
          "items": {
            "items": {
              "oneOf": [
                {
                  "$ref": "#/components/schemas/ReceivedDocument"
                },
                {
                  "$ref": "#/components/schemas/ProcessingDocument"
                },
                {
                  "$ref": "#/components/schemas/PendingApprovalDocument"
                },
                {
                  "$ref": "#/components/schemas/ApprovedDocument"
                },
                {
                  "$ref": "#/components/schemas/RejectedDocument"
                },
                {
                  "$ref": "#/components/schemas/FailedDocument"
                },
                {
                  "$ref": "#/components/schemas/SupersededDocument"
                }
              ],
              "discriminator": {
                "propertyName": "status",
                "mapping": {
                  "approved": "#/components/schemas/ApprovedDocument",
                  "failed": "#/components/schemas/FailedDocument",
                  "pending_approval": "#/components/schemas/PendingApprovalDocument",
                  "processing": "#/components/schemas/ProcessingDocument",
                  "received": "#/components/schemas/ReceivedDocument",
                  "rejected": "#/components/schemas/RejectedDocument",
                  "superseded": "#/components/schemas/SupersededDocument"
                }
              }
            },
            "type": "array",
            "title": "Items"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "items"
        ],
        "title": "DocumentListResponse"
      },
...
      "ErrorDetail": {
        "properties": {
          "code": {
            "type": "string",
            "maxLength": 80,
            "minLength": 1,
            "title": "Code"
          },
          "message": {
            "type": "string",
            "maxLength": 500,
            "minLength": 1,
            "title": "Message"
          },
          "issues": {
            "items": {
              "$ref": "#/components/schemas/ValidationIssue"
            },
            "type": "array",
            "title": "Issues"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "code",
          "message",
          "issues"
        ],
        "title": "ErrorDetail"
      },
      "ErrorResponse": {
        "properties": {
          "error": {
            "$ref": "#/components/schemas/ErrorDetail"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "error"
        ],
        "title": "ErrorResponse"
      },
...
      "RegisterDocumentRequest": {
        "properties": {
          "filename": {
            "type": "string",
            "pattern": "^[A-Za-z0-9][A-Za-z0-9._ -]{0,249}\\.[Pp][Dd][Ff]$",
            "title": "Filename"
          },
          "media_type": {
            "type": "string",
            "const": "application/pdf",
            "title": "Media Type"
          },
          "size_bytes": {
            "type": "integer",
            "maximum": 25000000.0,
            "minimum": 1.0,
            "title": "Size Bytes"
          },
          "sha256": {
            "type": "string",
            "pattern": "^[0-9a-f]{64}$",
            "title": "Sha256"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "filename",
          "media_type",
          "size_bytes",
          "sha256"
        ],
        "title": "RegisterDocumentRequest"
      },
...
    },
    "headers": {
      "AnalysisMode": {
        "description": "Modo de análise configurado, com valor sanitizado.",
        "schema": {
          "type": "string",
          "enum": [
            "synthetic_demo",
            "artifacts"
          ]
        }
      }
    }
  }
}